· GSL - F.16: For “in” parameters, pass cheaply copied types by value     and others by reference to const (WIP).
    For floats, doubles, std::uint8_t, std::uint16_t, std::uint32_t,
    std::uint64_t, enums, etc.

    ( http://isocpp.github.io/CppCoreGuidelines/
CppCoreGuidelines.html#a-
    namerf-ina-f16-for-in-parameters-pass-cheaply-copied-types-by-value-
and-
    others-by-reference-to-const )

+++ Source/SceneR/Content/ContentReader.hpp
@@ -397,34 +397,34 @@ inline Math::Matrix ContentReader::convert(const std::vector<json11::Json>& valu
 template<>
 inline Math::Quaternion ContentReader::convert(const std::vector<json11::Json>& values) const noexcept
 {
-    return { values[0].number_value()
-           , values[1].number_value()
-           , values[2].number_value()
-           , values[3].number_value() };
+    return { static_cast<float>(values[0].number_value())
+           , static_cast<float>(values[1].number_value())
+           , static_cast<float>(values[2].number_value())
+           , static_cast<float>(values[3].number_value()) };
 }
 
 template<>
 inline Math::Vector2 ContentReader::convert(const std::vector<json11::Json>& values) const noexcept
 {
-    return { values[0].number_value()
-           , values[1].number_value() };
+    return { static_cast<float>(values[0].number_value())
+           , static_cast<float>(values[1].number_value()) };
 }
 
 template<>
 inline Math::Vector3 ContentReader::convert(const std::vector<json11::Json>& values) const noexcept
 {
-    return { values[0].number_value()
-           , values[1].number_value()
-           , values[2].number_value()};
+    return { static_cast<float>(values[0].number_value())
+           , static_cast<float>(values[1].number_value())
+           , static_cast<float>(values[2].number_value()) };
 }
 
 template<>
 inline Math::Vector4 ContentReader::convert(const std::vector<json11::Json>& values) const noexcept
 {
-    return { values[0].number_value()
-           , values[1].number_value()
-           , values[2].number_value()
-           , values[3].number_value()};
+    return { static_cast<float>(values[0].number_value())
+           , static_cast<float>(values[1].number_value())
+           , static_cast<float>(values[2].number_value())
+           , static_cast<float>(values[3].number_value()) };
 }
 
 }}
